Move(#7): 공통 컴포넌트들 타입 파일위치 이동
