+++ src/types/components/commons/Card.ts
